# apps/web/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/rooms/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "AdminRoomsController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 메시지 운영 삭제 (tombstone)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/rooms/{id}": {
      "patch": {
        "operationId": "AdminRoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 상태 · active/closed/archived · 닫기·보관 시 reason",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/rooms": {
      "get": {
        "operationId": "RoomsController_listPublic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "공개 방 목록 조회",
        "tags": [
          "Rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 생성 (생성자=방장)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/mine": {
      "get": {
        "operationId": "RoomsController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 방 목록(멤버인 방)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/messages": {
      "get": {
        "operationId": "RoomsController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomMessageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 메시지 목록 조회",
        "tags": [
          "Rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 메시지 전송",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 상세 조회",
        "tags": [
          "Rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 설정 수정(방장)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/join": {
      "post": {
        "operationId": "RoomsController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "공개 방 입장",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/read": {
      "post": {
        "operationId": "RoomsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 메시지 읽음 처리",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/leave": {
      "post": {
        "operationId": "RoomsController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 퇴장",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/members/{userId}/kick": {
      "post": {
        "operationId": "RoomsController_kick",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "멤버 강퇴(방장)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/close": {
      "post": {
        "operationId": "RoomsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 닫기(방장만)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/transfer": {
      "post": {
        "operationId": "RoomsController_transfer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransfreRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방장 넘기기",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/members": {
      "get": {
        "operationId": "RoomsController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 멤버 목록(멤버만)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "RoomsController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "메시지 전체에서 삭제(작성자·방장)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/messages/{messageId}/hide": {
      "post": {
        "operationId": "RoomsController_hideMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "메시지 나에게서만 삭제(멤버)",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/theme": {
      "get": {
        "operationId": "RoomsController_getChatTheme",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomChatThemeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 채팅 테마 조회",
        "tags": [
          "Rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_updateChatTheme",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomChatThemeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomChatThemeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 채팅 테마 수정",
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}/messages/{messageId}/reactions": {
      "post": {
        "operationId": "RoomsController_toggleReaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleRoomMessageReactioinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "메시지 탭백 토글(멤버)",
        "tags": [
          "Rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Music Community API",
    "description": "Music Community API description",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "POST /auth/login 또는 /auth/register 후 발급받은 토큰을 사용하세요."
      }
    },
    "schemas": {
...
      "AuthorSnippetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nickname": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "nickname"
        ]
      },
...
      "UpdateAdminRoomDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "active",
              "closed",
              "archived"
            ],
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "방장 통지용 사유 · 닫기·보관 시 필수"
          }
        }
      },
...
      "RoomResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "topicTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "enum": [
              "public",
              "private",
              "invite"
            ],
            "type": "string"
          },
          "passwordHint": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "active",
              "closed",
              "archived"
            ],
            "type": "string"
          },
          "memberCount": {
            "type": "number"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/AuthorSnippetDto"
          },
          "lastMessageAt": {
            "type": "string",
            "nullable": true,
            "description": "`/rooms/mine`만"
          },
          "lastReadAt": {
            "type": "string"
          },
          "unread": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "topicTags",
          "visibility",
          "status",
          "memberCount",
          "ownerId",
          "createdAt",
          "updatedAt",
          "owner"
        ]
      },
...
      "RoomMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "recommendation",
              "saved_card",
              "lyric_quote",
              "system"
            ]
          },
          "body": {
            "type": "string",
            "nullable": true
          },
          "recommendationId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          },
          "deletedByOwner": {
            "type": "boolean"
          },
          "deletedById": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "sender": {
            "$ref": "#/components/schemas/RoomMessageSenderDto"
          },
          "recommendation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomMessageRecommendationDto"
              }
            ]
          },
          "savedCard": {
            "type": "object",
            "nullable": true
          },
          "lyricStartSec": {
            "type": "number",
            "nullable": true
          },
          "lyricEndSec": {
            "type": "number",
            "nullable": true
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomMessageReactionDto"
            }
          },
          "replyTo": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "roomId",
          "senderId",
          "type",
          "createdAt",
          "deletedByOwner",
          "sender"
        ]
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "topicTags": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 20
            }
          },
          "name": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 200
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "invite"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 64
          },
          "passwordHint": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "name"
        ]
      },
      "JoinRoomDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "topicTags": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 20
            }
          },
          "name": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "invite"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 64
          },
          "passwordHint": {
            "type": "string",
            "nullable": true,
            "maxLength": 40
          }
        }
      },
      "TransfreRoomDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "새 방장 — 이미 그 방 멤버",
            "format": "uuid"
          }
        },
        "required": [
          "userId"
        ]
      },
      "CreateRoomMessageDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "text",
              "recommendation",
              "saved_card",
              "lyric_quote"
            ]
          },
          "body": {
            "type": "string",
            "maxLength": 2000
          },
          "recommendationId": {
            "type": "string",
            "format": "uuid"
          },
          "savedCardId": {
            "type": "string",
            "format": "uuid"
          },
          "lyricStartSec": {
            "type": "number",
            "description": "선택 — 비우면 검증 스킵 (ValidateIf만 쓰면 undefined에도 IsInt 적용됨)",
            "minimum": 0
          },
          "lyricEndSec": {
            "type": "number",
            "minimum": 0
          },
          "replyToId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "type"
        ]
      },
      "RoomChatThemeResponseDto": {
        "type": "object",
        "properties": {
          "presetId": {
            "type": "string"
          },
          "backgroundUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "presetId"
        ]
      },
      "UpdateRoomChatThemeDto": {
        "type": "object",
        "properties": {
          "presetId": {
            "enum": [
              "lp-bar",
              "cream-paper",
              "midnight",
              "fan-pink"
            ],
            "type": "string"
          },
          "backgroundUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 2048
          }
        }
      },
      "ToggleRoomMessageReactioinDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          }
        },
        "required": [
          "emoji"
        ]
      },
...
    }
  }
}